add class to select options
We will need to know which tabs to activate when a discussion is chosen based on whether or not the discussion is threaded. Also which tab to activate. These changes work in conjunction with others I will be making to threadz.php in the success function triggered by loading ajax.php